% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{analysisServer}
\alias{analysisServer}
\title{Analysis Server}
\usage{
analysisServer(id, reactiveEndpoint)
}
\arguments{
\item{id}{mod ID}

\item{reactiveEndpoint}{reactive element with the type of endpoint}
}
\description{
Analysis Server
}
