% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastt-package.R
\docType{package}
\name{beastt-package}
\alias{beastt-package}
\alias{beastt}
\title{The 'beastt' package.}
\description{
Inverse Probablity Weighted Baysian Dynamic Barrowing
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gsk-biostatistics.github.io/beastt/}
  \item \url{https://github.com/GSK-Biostatistics/beastt}
  \item Report bugs at \url{https://github.com/GSK-Biostatistics/beastt/issues}
}

}
\author{
\strong{Maintainer}: Christina Fillmore \email{christina.e.fillmore@gsk.com} (\href{https://orcid.org/0000-0003-0595-2302}{ORCID})

Authors:
\itemize{
  \item Ben Arancibia \email{benjamin.c.arancibia@gsk.com}
  \item Nate Bean \email{nate.w.bean@gsk.com} (\href{https://orcid.org/0000-0001-9946-0119}{ORCID})
}

Other contributors:
\itemize{
  \item GlaxoSmithKline Research & Development Limited [copyright holder, funder]
}

}
