% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-sim.R
\name{calc_cond_binary}
\alias{calc_cond_binary}
\title{Calculate Conditional Drift for Binary}
\usage{
calc_cond_binary(population, glm, marg_drift, marg_trt_eff)
}
\arguments{
\item{population}{Large population dataset, where the columns of the dataset
correspond to the covariates for the logistic regression}

\item{glm}{logistic regression model object}

\item{marg_drift}{vector of marginal drift values}

\item{marg_trt_eff}{vector of marginal treatment effect values}
}
\value{
tibble of all combinations of the marginal drift and treatment
effect. For each row the conditional drift and treatment effect has been
calculated as well as the true control response rate and true treatment
effect.
}
\description{
In order to properly simulate values to test inverse probability weighting,
we need to convert the marginal drift and treatment effect to effects
conditioned on the predictive covariates we will use in the IPW weighting.
}
