% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{calc_post_beta}
\alias{calc_post_beta}
\title{Calculate Posterior Beta}
\usage{
calc_post_beta(internal_data, response, prior)
}
\arguments{
\item{internal_data}{This can either be a propensity score object or a tibble
of the internal data.}

\item{response}{Name of response variable}

\item{prior}{distributional object, if you would like a mixture distribution}
}
\value{
distributional object
}
\description{
Calculate Posterior Beta
}
