% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prior.R
\name{calc_power_prior}
\alias{calc_power_prior}
\title{Calculate Power Prior}
\usage{
calc_power_prior(
  dist = c("binom", "norm"),
  hyperparameter = NULL,
  weighted_obj,
  response_var
)
}
\arguments{
\item{dist}{Name of the distribution of the external data response}

\item{hyperparameter}{Named vector of the hyperparameters for a given
distribution. If you are using an improper prior then value can be left as
\code{NULL}}

\item{weighted_obj}{A \code{prop_scr_obj} created by calling \code{calc_prop_scr()}}

\item{response_var}{Name of response variable}
}
\value{
\code{power_prior} object
}
\description{
Function to calculate a power prior given a distribution, hyperparameters, a
weighted object and a response variable. The weighted object will typically
be a \code{prop_scr_obj} created by calling \code{create_prop_scr()}
}
