% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prior.R
\name{calc_power_prior_norm}
\alias{calc_power_prior_norm}
\title{Calculate Power Prior Normal}
\usage{
calc_power_prior_norm(
  weighted_obj,
  response,
  prior = NULL,
  external_control_sd = NULL
)
}
\arguments{
\item{weighted_obj}{A \code{prop_scr_obj} created by calling \code{create_prop_scr()}}

\item{response}{Name of response variable}

\item{prior}{either \code{NULL} or a normal distributional object that is the
initial prior for the control mean before the external control data are observed}

\item{external_control_sd}{Standard deviation of external control response data if
assumed known. It can be left as \code{NULL} if assumed unknown}
}
\value{
normal power prior object
}
\description{
Calculate Power Prior Normal
}
\seealso{
Other power prior: 
\code{\link{calc_power_prior_beta}()}
}
\concept{power prior}
