% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prior.R
\name{calc_power_prior_weibull}
\alias{calc_power_prior_weibull}
\title{Calculate Power Prior Weibull}
\usage{
calc_power_prior_weibull(
  external_data,
  response,
  event,
  intercept,
  shape,
  approximation = c("Laplace", "MCMC"),
  ...
)
}
\arguments{
\item{external_data}{This can either be a \code{prop_scr_obj} created by calling
\code{create_prop_scr()} or a tibble of the external data. If it is just a
tibble the weights will be assumed to be 1. Only the external data for the
arm(s) of interest should be included in this object (e.g., external
control data if creating a power prior for the control mean)}

\item{response}{Name of response variable}

\item{event}{Name of event variable}

\item{intercept}{Normal distributional object that is the initial prior for the
log-inverse-scale parameter}

\item{shape}{Integer value that is the scale of the half-normal prior
for the shape parameter}

\item{approximation}{Type of approximation to be used. Either La Place or
MCMC, La Place is used by default because it is faster than MCMC.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
Multivariate Normal Distributional Object
}
\description{
Calculate a (potentially inverse probability weighted) normal
power prior using external data.
}
