% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{calc_std_mean_diff}
\alias{calc_std_mean_diff}
\title{Calculate Absolute Standardized Mean Differences}
\usage{
calc_std_mean_diff(prop_scr_obj, treatment_var, ...)
}
\arguments{
\item{prop_scr_obj}{Propensity score object}

\item{treatment_var}{Treatment variable}

\item{...}{Any additional variables}
}
\value{
Tibble of the calculated standardized mean differences for the
additional variables
}
\description{
Calculate the unadjusted and IPW-adjusted absolute standardized
mean differences for each covariate.
}
