% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{calc_weibull_post}
\alias{calc_weibull_post}
\title{Calculate Weibull Posterior}
\usage{
calc_weibull_post(internal_data, response, event, prior, analysis_time, ...)
}
\arguments{
\item{internal_data}{This can either be a propensity score object or a tibble
of the internal data.}

\item{response}{Name of response variable}

\item{event}{Name of event/censoring variable}

\item{prior}{A distributional object corresponding to a multivariate normal
distribution or a mixture of 2 multivariate normals}

\item{analysis_time}{Vector of time(s) when survival probabilities will be
calculated}

\item{...}{rstan sampling option. This override any default options. Fro more
information see \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}
}
\value{
stan posterior object
}
\description{
Calculate a posterior distribution for time to event data with a
Weibull Likelihood. Only the relevant treatment arms from the internal
dataset should be read in (e.g., only the control arm if constructing a
posterior distribution for the control response rate). To calcualate
treatment difference calculate posterior or both arms then subtract.
}
