% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_binary_df}
\alias{ex_binary_df}
\title{External Binary Data for Propensity Score Matching}
\format{
\subsection{\code{ex_binary_df}}{

A data frame with 150 rows and 6 columns:
\describe{
\item{subjid}{Unique subject ID}
\item{cov1}{Covariate 1, which is normally distributed around 65 with an sd of 15}
\item{cov2}{Covariate 2, which is binary with about a 30\% response rate}
\item{cov3}{Covariate 3, which is binary with about a 40\% response rate}
\item{cov4}{Covariate 4, which is binary with about a 50\% response rate}
\item{y}{Response, which is binary}
}
}
}
\usage{
ex_binary_df
}
\description{
This is a simulated dataset used to illustrate bayesian dynamic borrowing in
the binary case.
}
\keyword{datasets}
