% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{int_binary_df}
\alias{int_binary_df}
\title{Internal Binary Data for Propensity Score Matching}
\format{
\subsection{\code{int_binary_df}}{

A data frame with 150 rows and 6 columns:
\describe{
\item{subjid}{Unique subject ID}
\item{cov1}{Covariate 1, which is normally distributed around 62 with an sd of 5}
\item{cov2}{Covariate 2, which is binary with about a 60\% response rate}
\item{cov3}{Covariate 3, which is binary with about a 40\% response rate}
\item{cov4}{Covariate 4, which is binary with about a 30\% response rate}
\item{trt}{Treatment indicator, where 0 = control and 1 = active treatment}
\item{y}{Response, which is binary}
}
}
}
\usage{
int_binary_df
}
\description{
This is a simulated dataset used to illustrate bayesian dynamic borrowing in
the binary case.
}
\keyword{datasets}
