% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{is_prop_scr}
\alias{is_prop_scr}
\title{Test If Propensity Score Object}
\usage{
is_prop_scr(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Boolean
}
\description{
Test If Propensity Score Object
}
\examples{
library(dplyr)
x <- calc_prop_scr(internal_df = filter(int_norm_df, trt == 0),
                       external_df = ex_norm_df,
                       id_col = subjid,
                       model = ~ cov1 + cov2 + cov3 + cov4)
is_prop_scr(x)

}
