% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot Distribution}
\usage{
plot_dist(...)
}
\arguments{
\item{...}{Distributional object(s) to plot. If there are multiple objects you can name them.}
}
\value{
ggplot object that is the density of the provided distribution
}
\description{
Plot Distribution
}
