% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{plotServer}
\alias{plotServer}
\title{Plot Inputs server}
\usage{
plotServer(id, input_list, robust)
}
\arguments{
\item{id}{mod id}

\item{input_list}{list of inputs}

\item{robust}{robustify}
}
\description{
Plot Inputs server
}
