% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{pre_post_summary}
\alias{pre_post_summary}
\title{Create a Pre-and-Post-Weighting Summary Table}
\usage{
pre_post_summary(
  df,
  vars,
  internal,
  trt,
  internal_df,
  external_df,
  id_col,
  model,
  ...
)
}
\arguments{
\item{df}{Combined dataset with one row per subject and all the
variables (subgroups) to be displayed in the summary table}

\item{vars}{Variables (subgroups) to be displayed in the summary table}

\item{internal}{Indicator variable identifying internal vs external subjects}

\item{trt}{Treatment variable, should be set to "Control" if borrowing from
external controls}

\item{internal_df}{Internal dataset with one row per subject and all the
variables needed to run the model}

\item{external_df}{External dataset with one row per subject and all the
variables needed to run the model}

\item{id_col}{Name of the column in both datasets used to identify each
subject. It must be the same across datasets}

\item{model}{Model used to calculate propensity scores}

\item{...}{Optional arguments}
}
\value{
\code{tibble} object, with the adjusted and unadjusted
}
\description{
Create a Pre-and-Post-Weighting Summary Table
}
\examples{
df <- data.frame(
  id_col = 1:40,
  cov1 = rep(c("a", "b", "c", "d"), 10),
  cov2 = rep(c("e", "f", "g", "h"), each = 10),
  internal = sample(c(0,1), 40, replace = TRUE),
  trt = rep(c("Active", "Control"), each = 20)
)
model <- as.formula("~cov1 + cov2")
internal_df <- df |>
  dplyr::filter(internal == 1)
external_df <- df |>
  dplyr::filter(internal == 0)
pre_post_summary(df, vars = c("cov1", "cov2"), df$internal, trt = df$trt,
internal_df, external_df, model = model, id_col = id_col)
}
