% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{prop_scr_cloud}
\alias{prop_scr_cloud}
\title{Propensity Score Cloud Plot}
\usage{
prop_scr_cloud(x, trimmed_prop_scr = NULL)
}
\arguments{
\item{x}{a \code{prop_scr} obj}

\item{trimmed_prop_scr}{a trimmed \code{prop_scr} obj}
}
\value{
ggplot object
}
\description{
Propensity Score Cloud Plot
}
