% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{prop_scr_hist}
\alias{prop_scr_hist}
\title{Histogram of the Propensity Score Object}
\usage{
prop_scr_hist(
  x,
  variable = c("propensity score", "ps", "inverse probability weight", "ipw"),
  ...
)
}
\arguments{
\item{x}{Propensity score object}

\item{variable}{Variable to plot. It must be either a propensity score
("ps" or "propensity score") or inverse probability weight ("ipw" or
"inverse probability weight")}

\item{...}{Optional arguments for \code{geom_histogram}}
}
\value{
ggplot object
}
\description{
Plot overlapping histograms of the propensity scores for both
the internal and external participants, or plot external IPWs.
}
