% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{rescale}
\alias{rescale}
\title{Rescale a prop_scr object}
\usage{
rescale(x, n = NULL, scale_factor = NULL)
}
\arguments{
\item{x}{a \code{prop_scr} obj}

\item{n}{Desired sample size that the external data effectively contribute to
the analysis of the internal trial data. This will be used to scale the
external weights if no scale factor is given}

\item{scale_factor}{Value to multiple all weights by.}
}
\value{
a rescaled propensity score
}
\description{
Rescale a prop_scr object
}
\examples{
library(dplyr)
ps_obj <- calc_prop_scr(internal_df = filter(int_binary_df, trt == 0),
                       external_df = ex_binary_df,
                       id_col = subjid,
                       model = ~ cov1 + cov2 + cov3 + cov4)
# propensity score objects can be rescaled to equal a desired outcome weight
rescale(ps_obj, n = 75)

# Or by a predetermined factor
rescale(ps_obj, scale_factor = 1.5)

}
