% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robustify_mvnorm}
\alias{robustify_mvnorm}
\title{Robustify Multivariate Normal Distributions}
\usage{
robustify_mvnorm(prior, n, weights = c(0.5, 0.5))
}
\arguments{
\item{prior}{Multivariate Normal distributional object}

\item{n}{Number of theoretical participants}

\item{weights}{Vector of weights, where the first number corresponds to the
informative component and the second is the vague}
}
\value{
mixture distribution
}
\description{
Adds vague normal component, where the level of vagueness is controlled by
the \code{n} parameter
}
\details{
In cases with a normal endpoint, a robust mixture prior can be created by
adding a vague normal component to any normal prior with mean \eqn{\theta}
and variance \eqn{\sigma^2}.The vague component is calculated to have the
same mean \eqn{\theta} and variance equal to \eqn{\sigma^2 \times n}, where
\code{n} is the specified number of theoretical participants. If robustifying a normal
power prior that was calculated from external control data and \code{n} is defined as
the number of external control participants, and the vague component would
then correspond to one external control participant's worth of data.
}
\examples{
library(distributional)
robustify_mvnorm(
      dist_multivariate_normal(mu = list(c(1, 0)), sigma = list(c(10, 5))),
       n = 15)
}
