% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte-sim.R
\name{simulate_accrual}
\alias{simulate_accrual}
\title{Simulate participant accrual times}
\usage{
simulate_accrual(n, accrual_periods, accrual_props)
}
\arguments{
\item{n}{Number of participants}

\item{accrual_periods}{Vector of right endpoints defining the time periods of
accrual, e.g., c(6,8) defines [0,6), \link{6,8}.}

\item{accrual_props}{Vector indicating the proportion of participants that are
expected to be enrolled during each of the defined accrual periods. Must sum
to 1.}
}
\value{
Vector of accrual times corresponding to when each participant enters
the study
}
\description{
Simulate participant accrual times
}
\examples{
at <- simulate_accrual(n = 100000, accrual_periods = c(6, 8), accrual_props = c(.5, .5))
hist(at, breaks = 100, main = "Histogram of Enrollment Times", xlab = "Enrollment Time")
}
