% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tte-sim.R
\name{sim_weib_ph}
\alias{sim_weib_ph}
\title{Simulate event times for each participant from a Weibull proportional hazards
regression model}
\usage{
sim_weib_ph(
  weibull_ph_mod,
  samp_df,
  cond_drift = 0,
  cond_trt_effect = 0
)
}
\arguments{
\item{weibull_ph_mod}{\code{survreg} object corresponding to a Weibull
proportional hazards model fit using the external data}

\item{samp_df}{Data frame of covariates corresponding to the sample arm
(control or treated) for which event times should be simulated. The column
names should correspond to the covariate names in the \code{survreg} object.}

\item{cond_drift}{Optional value of the conditional drift by which the
intercept in the Weibull proportional hazards regression model should be
increased/decreased to incorporate the impact of unmeasurable sources of
drift. Default is 0.}

\item{cond_trt_effect}{Optional value of the conditional treatment effect by
which the intercept in the Weibull proportional hazards regression model
should be increased/decreased if simulating event data for a treated arm.
Default is 0.}
}
\value{
Vector of simulated event times from a Weibull proportional hazards
regression model
}
\description{
Simulate event times for each participant from a Weibull proportional hazards
regression model
}
\examples{
library(dplyr)
library(survival)
# Model "true" regression coefficients and shape parameter using the external data
weibull_ph_mod <- survreg(Surv(y, event) ~ cov1 + cov2 + cov3 + cov4, data = ex_tte_df,
                          dist = "weibull")

# Sample covariates for internal control arm via bootstrap from external data
samp_int_ctrl <- bootstrap_cov(ex_tte_df, n = 100) |>
  select(c(cov1, cov2, cov3, cov4))     # keep only covariate columns
tte_dat <- sim_weib_ph(weibull_ph_mod, samp_df = samp_int_ctrl)
}
