% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_simulation.R
\name{simulationServer}
\alias{simulationServer}
\title{Simulation Server}
\usage{
simulationServer(id, input_list)
}
\arguments{
\item{id}{mod ID}

\item{input_list}{input from UI}
}
\description{
Simulation Server
}
