% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{trim}
\alias{trim}
\title{Trim a prop_scr object}
\usage{
trim(x, low = NULL, high = NULL)
}
\arguments{
\item{x}{a \code{prop_scr} obj}

\item{low}{Low cut-off so all values will be greater than. If left \code{NULL} no
lower bound will be used}

\item{high}{High cut-off so all values will be less than. If left \code{NULL} no
upper bound will be used}
}
\value{
a trimmed propensity score
}
\description{
Trim a prop_scr object
}
\examples{
library(dplyr)
ps_obj <- calc_prop_scr(internal_df = filter(int_binary_df, trt == 0),
                       external_df = ex_binary_df,
                       id_col = subjid,
                       model = ~ cov1 + cov2 + cov3 + cov4)
trim(ps_obj, low = 0.3, high = 0.7)

}
