% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_code.R
\name{write_code}
\alias{write_code}
\title{Write Code function}
\usage{
write_code(simulation, endpoint, selections)
}
\arguments{
\item{simulation}{A purpose, either "Simulation" or "Analysis"}

\item{endpoint}{An endpoint type, one of "Binary", "Normal" or "Survival"}

\item{selections}{input from UI}
}
\description{
Write Code function
}
