% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_analysis.R
\name{analysisServer}
\alias{analysisServer}
\title{Analysis Server}
\usage{
analysisServer(id, input_list)
}
\arguments{
\item{id}{mod ID}

\item{input_list}{input from UI}
}
\description{
Analysis Server
}
