% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beastt-package.R
\docType{_PACKAGE}
\name{beastt-package}
\alias{beastt-package}
\alias{beastt}
\title{The 'beastt' package.}
\description{
Inverse Probablity Weighted Baysian Dynamic Barrowing
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.3. https://mc-stan.org
}
