% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{binary_sim_df}
\alias{binary_sim_df}
\title{Binary Simulation Data}
\format{
\subsection{\code{binary_sim_df} A data frame with 255 rows and 6 columns:}{

\describe{
\item{population}{Different population imbalances}
\item{marg_trt_eff}{marginal treatment effect}
\item{true_control_RR}{True control response rate on the marginal scale}
\item{reject_H0_yes}{probability of rejecting the null hypothesis in the case with borrowing}
\item{no_borrowing_reject_H0_yes}{probability of rejecting the null hypothesis without borrowing}
\item{pwr_prior}{vector of power priors distributional objects}
}
}
}
\usage{
binary_sim_df
}
\description{
This is a simulated dataset used to illustrate Bayesian dynamic borrowing
simulations. This was generated based on binary simulation template. In this
simulation only population and marginal treatment effect are being varied
}
\keyword{datasets}
