% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-sim.R
\name{bootstrap_cov}
\alias{bootstrap_cov}
\title{Bootstrap Covariate Data}
\usage{
bootstrap_cov(
  external_dat,
  n,
  imbal_var = NULL,
  imbal_prop = NULL,
  ref_val = 0
)
}
\arguments{
\item{external_dat}{External Data to bootstrap}

\item{n}{number of rows in the output dataset}

\item{imbal_var}{optional variable to to create imbalance with. If left
\code{NULL} the distribution of the covariates will match the distribution in
the external dataset. The imbalance variable must be binary.}

\item{imbal_prop}{optional imbalance proportion. Needed if there is an
imbalance variable. This controls the proportion of subjects with the
reference value. This can either be a vector of proportions or a single
proportion.}

\item{ref_val}{}
}
\value{
dataframe with the same number of columns as the reference
dataframe, but n number of rows
}
\description{
Bootstrap Covariate Data
}
\examples{
bootstrap_cov(external_dat, n = 100000)
}
