% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{calc_post_norm}
\alias{calc_post_norm}
\title{Calculate Posterior Normal}
\usage{
calc_post_norm(internal_data, response, prior, internal_control_sd = NULL)
}
\arguments{
\item{internal_data}{This can either be a propensity score object or a tibble
of the internal data.}

\item{response}{Name of response variable}

\item{prior}{distributional object, if you would like a mixture distribution}

\item{internal_control_sd}{Assumed known SD of the internal control}
}
\value{
distributional object
}
\description{
Calculate Posterior Normal
}
