% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prior.R
\name{calc_power_prior}
\alias{calc_power_prior}
\title{Calculate Power Prior}
\usage{
calc_power_prior(prior, weighted_obj, response, ...)
}
\arguments{
\item{prior}{a {distributional} object that is the prior of the external data}

\item{weighted_obj}{A \code{prop_scr_obj} created by calling \code{calc_prop_scr()}}

\item{response}{Name of response variable}

\item{...}{additional parameters needed depending on the type of power prior
you are trying to calculate TODO: Make this a bulleted list with the
additional parameters for each distirbution}
}
\value{
\code{power_prior} object
}
\description{
Function to calculate a power prior given a distribution, hyperparameters, a
weighted object and a response variable. The weighted object will typically
be a \code{prop_scr_obj} created by calling \code{create_prop_scr()}
}
