% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwr_prior.R
\name{calc_power_prior_beta}
\alias{calc_power_prior_beta}
\title{Calculate Power Prior Beta}
\usage{
calc_power_prior_beta(prior, weighted_obj, response)
}
\arguments{
\item{prior}{a beta distributional object that is the prior of the external data}

\item{weighted_obj}{A \code{prop_scr_obj} created by calling \code{create_prop_scr()}}

\item{response}{Name of response variable}
}
\value{
beta power prior object
}
\description{
Calculate Power Prior Beta
}
\seealso{
Other power prior: 
\code{\link{calc_power_prior_norm}()}
}
\concept{power prior}
