% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{calc_prop_scr}
\alias{calc_prop_scr}
\title{Create a Proprensity Score Object}
\usage{
calc_prop_scr(internal_df, external_df, id_col, model, ...)
}
\arguments{
\item{internal_df}{Internal dataset with one row per subject and all the
variables needed to run the model}

\item{external_df}{External dataset with one row per subject and all the
variables needed to run the model}

\item{id_col}{Name of the column in both datasets used to identify each
subject. It must be the same across datasets}

\item{model}{Model used to calculate propensity scores}

\item{...}{Optional arguments}
}
\value{
\code{prop_scr_obj} object, with the internal and the external data and
the propensity score and inverse probability weight calculated for each
subject.
}
\description{
Create a Proprensity Score Object
}
