% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{calc_std_mean_diff}
\alias{calc_std_mean_diff}
\title{Title}
\usage{
calc_std_mean_diff(prop_scr_obj, treatment_var, ...)
}
\arguments{
\item{prop_scr_obj}{propenstiy score object}

\item{treatment_var}{treatment variable}

\item{...}{any additional variables}
}
\value{
tibble of the calculated standardised mean differences for the additional variables
}
\description{
Title
}
