% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_norm_df}
\alias{ex_norm_df}
\title{External Normal Data for Propensity Score Matching}
\format{
\subsection{\code{ex_norm_df}}{

A data frame with 150 rows and 6 columns:
\describe{
\item{subjid}{Unique subject ID}
\item{cov1}{Covariate 1, which is normally distributed around 50 with an sd of 10}
\item{cov2}{Covariate 2, which is binary with about a 20\% response rate}
\item{cov3}{Covariate 3, which is binary with about a 60\% response rate}
\item{cov4}{Covariate 4, which is binary with about a 30\% response rate}
\item{y}{Response, which is normally distributed around 45 with an sd of 5}
}
}
}
\usage{
ex_norm_df
}
\description{
This is a simulated dataset used to illustrate bayesian dynamic borrowing in
the normal case.
}
\keyword{datasets}
