% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{int_norm_df}
\alias{int_norm_df}
\title{Internal Normal Data for Propensity Score Matching}
\format{
\subsection{\code{int_norm_df}}{

A data frame with 120 rows and 7 columns:
\describe{
\item{subjid}{Unique subject ID}
\item{cov1}{Covariate 1, which is normally distributed around 55 with an sd of 8}
\item{cov2}{Covariate 2, which is binary with about a 30\% response rate}
\item{cov3}{Covariate 3, which is binary with about a 50\% response rate}
\item{cov4}{Covariate 4, which is binary with about a 30\% response rate}
\item{trt}{Treatment indicator, where 0 = control and 1 = active treatment}
\item{y}{Response, which is normally distributed around 50 with an sd of 3}
}
}
}
\usage{
int_norm_df
}
\description{
This is a simulated dataset used to illustrate bayesian dynamic borrowing in
the normal case.
}
\keyword{datasets}
