% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-sim.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{value(s) to take the inverse logit of}
}
\value{
vector of numbers between 0 and 1
}
\description{
Inverse Logit
}
\details{
This function is a short hand to exp(x)/(1+exp(x))
}
\examples{
inv_logit(0.5)

}
