% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_dist}
\alias{plot_dist}
\title{Plot Distribution}
\usage{
plot_dist(...)
}
\arguments{
\item{...}{Distributional object(s) to plot. When passing multiple objects
naming them will change the labels in the plot, else they will use the
distributional format}
}
\value{
ggplot object that is the density of the provided distribution
}
\description{
Plot Distribution
}
\examples{
library(distributional)
plot_dist(dist_normal(0, 1))
#Plotting Multiple
plot_dist(dist_normal(0, 1), dist_normal(10, 5))
plot_dist('Prior' = dist_normal(0, 1), 'Posterior' = dist_normal(10, 5))
}
