% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_dist}
\alias{plot_dist}
\title{Title}
\usage{
plot_dist(distribution)
}
\arguments{
\item{distribution}{Distributional object to plot}
}
\value{
ggplot object that is the density of the provided distribution
}
\description{
Title
}
