% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{prop_scr_love}
\alias{prop_scr_love}
\title{Density of the Propensity Score Object}
\usage{
prop_scr_love(x, ...)
}
\arguments{
\item{x}{Propensity score object}

\item{...}{Optional options for \code{geom_point}}
}
\value{
ggplot object
}
\description{
Density of the Propensity Score Object
}
