% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{robustify_norm}
\alias{robustify_norm}
\title{Robustify Normal Distributions}
\usage{
robustify_norm(prior, n, weights = c(0.5, 0.5))
}
\arguments{
\item{prior}{Normal distributional object}

\item{n}{Number of theoretical participants}

\item{weights}{Vector of weights, where the first number corresponds to the
informative component and the second is the vague}
}
\value{
mixture distribution
}
\description{
Adds vague normal component, where the level of vagueness is controlled by
the \code{n} parameter
}
\examples{
library(distributional)
dist_normal(dist_norm(0,1), n = 15)
}
