% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps.R
\name{trim}
\alias{trim}
\title{Trim a \code{prop_scr} object}
\usage{
trim(x, low = NULL, high = NULL, quantile = FALSE)
}
\arguments{
\item{x}{a \code{prop_scr} object}

\item{low}{Low cut-off such that all participants with propensity scores (or
quantiles if \code{quantile = TRUE}) less than this value are removed. If left
\code{NULL} no lower bound will be used}

\item{high}{High cut-off such that all participants with propensity scores
(or quantiles if \code{quantile = TRUE}) greater than this value are removed. If
left \code{NULL} no upper bound will be used}

\item{quantile}{True/False value to determine if the cut-off values are based
directly one the propensity scores or their quantiles.}
}
\value{
a \code{prop_scr} object with a trimmed propensity score distribution
}
\description{
Trim a \code{prop_scr} object
}
\examples{
library(dplyr)
ps_obj <- calc_prop_scr(internal_df = filter(int_binary_df, trt == 0),
                       external_df = ex_binary_df,
                       id_col = subjid,
                       model = ~ cov1 + cov2 + cov3 + cov4)
trim(ps_obj, low = 0.3, high = 0.7)

}
